requirements.txt
langchain
langchain-core
langchain_community
langchain-openai
openai

python-dotenv
python-jose
passlib
pyJWT
pydantic
pydantic_core
pydantic[email]
bcrypt==4.3.0

motor


langchain_text_splitters
fastapi
requests
uvicorn
python-multipart
pypdf
faiss-cpu